Move PropertyTester into new module abstract_test_bases.py
